Sentry Power
Got a very rough version of the sentry power working. Press 'F' when
looking at an enemy who has a weapon you can control. When F is pressed,
gun detaches from enemy and starts routines to search for nearest
enemies that are in line of sight.

NEED TO TEST:
- is line of sight actually working?
- how fast does the weapon track? Need it to have some sort of slow
turning so enemy has a chance to move away from it

NEED TO DO:
- Add sound effects and muzzle flash. Basically, make it look decent
when firing.
- Add some sort of rotation delay for the gun so it doesn't instantly
lock on and fire at a target all the time.
- Some kind of punishment for using the skill. Maybe a cooldown time or
a massive stamina drain.